pyproject.toml
[project]
name = "py-html-renderer"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "pchauv" }
]
dependencies = [
    "babel>=2.18.0",
    "jinja2>=3.1.6",
    "libsass>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/pchauv"
Repository = "https://github.com/pchauv/py-html-renderer"
Issues = "https://github.com/pchauv/py-html-renderer/issues"

[dependency-groups]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/py_html_renderer"]

[tool.uv]
managed = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", "F", "W", # Erreurs, Pyflakes, Warnings
    "I",           # Isort (tri des imports)
    "N",           # Pep8-naming
    "UP",          # Pyupgrade (syntaxe moderne)
    "B",           # Bugbear (erreurs communes)
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
